Content_Extraction/requirements.txt
# Core dependencies for SCORM content extraction
PyPDF2>=3.0.0
moviepy>=1.0.3

# Image processing and OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# HTML parsing and processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTTP requests and Elasticsearch integration
requests>=2.31.0
elasticsearch>=8.0.0

# Azure Speech Services for video/audio transcription
azure-cognitiveservices-speech>=1.31.0

# Web interface dependencies
Flask>=2.3.0
Werkzeug>=2.3.0

# Additional utilities
python-dotenv>=0.10.0 